Leetcode/src/test/java/problem/tree/Lc695maxAreaOfIslandTest.java
package problem.tree;

import org.junit.jupiter.api.Test;
import problem.array.parse.ArrayParse;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Lc695maxAreaOfIslandTest {

    @Test
    void maxAreaOfIsland() {
        assertEquals(4, new Lc695maxAreaOfIsland().maxAreaOfIsland(ArrayParse.parse2DIntArray("[[1,1,0,0,0],[1,1,0,0,0],[0,0,0,1,1],[0,0,0,1,1]]")));
        assertEquals(6, new Lc695maxAreaOfIsland().maxAreaOfIsland(ArrayParse.parse2DIntArray("[[0,0,1,0,0,0,0,1,0,0,0,0,0],[0,0,0,0,0,0,0,1,1,1,0,0,0],[0,1,1,0,1,0,0,0,0,0,0,0,0],[0,1,0,0,1,1,0,0,1,0,1,0,0],[0,1,0,0,1,1,0,0,1,1,1,0,0],[0,0,0,0,0,0,0,0,0,0,1,0,0],[0,0,0,0,0,0,0,1,1,1,0,0,0],[0,0,0,0,0,0,0,1,1,0,0,0,0]]")));
    }
}